Getting to 100% coverage in `webutil/request_option.go`.
It's unclear if this is worth it. It was accomplished by replacing
`b := new(bytes.Buffer)` with a call to a mutable default
`makeBytesBuffer`. In order to enable this to stand in and be mocked,
we also introduced a stand-in interface `bytesWriter` and two
helpers `setMakeBytesBuffer` / `restoreMakeBytesBuffer` for replacing
and restoring the mutable default.

It is essentially **required** to get test coverage because none
of the lines in `populateFormData` **can** fail with a "fresh"
`*multipart.Writer` that wraps a `new(bytes.Buffer)`.