Getting to 100% coverage in `webutil/local_ip.go`.
It's unclear if this is worth it. It was accomplished by replacing
`net.InterfaceAddrs` with a a mutable default `getInterfaceAddrs`. In
order to enable this to stand in and be mocked, we also introduced two
helpers `setGetInterfaceAddrs` / `restoreGetInterfaceAddrs` for
replacing and restoring the mutable default.

It is essentially **required** to get test coverage because the
returned interface addresses are completely system dependent, so the
"fallthrough to loopback" or "error getting addresses" cases are
(effectively) impossible to trigger.